requirements.txt
Django==1.11.1
django-debug-toolbar==1.8
gunicorn==19.7.1
jsonschema
django-ipware
django-extensions